==> Scripts_invivo_Mouse/Step0_2_BRCA_ClusterAnnotation.Rmd <==
---
title: "Step1_2_BRCA_ClusterAnnotation"
author: "Melissa"
date: "6/22/2021"
output: html_document
---

```{r}
mainDir <- here::here() # the MainDir should be the root of Tumor_BRCA_P53:
source(knitr::purl(file.path(mainDir,"Scripts","global_var_current.Rmd"), quiet=TRUE))
source(knitr::purl(file.path(mainDir,"Scripts","scRNAseq_Functions.Rmd"), quiet=TRUE))
source(knitr::purl(file.path(mainDir,"Scripts","global_variables.Rmd"), quiet=TRUE))
```

Load the BRCA seurat Object which has been saved adding the annotation:
```{r}
brca=qs::qread(paste0(RDatadir, "/BRCA_SCT_1k10k.qs"))
```

Once you load this brca qs file, Go directly to line 110 if you want to generate UMAPS and see the coarse grained annotation


#----------------------------------------------SKIP THIS PART ---------------------------------------------------------#
#----------------------------------------------------------------------------------------------------------------------#
#-------------------------------- In vivo Data Cluster Annotation ------------------------------------#
-------------------------------------------------------------------------------------------------------

Pick up the the canonical marker genes from the expression matrix:
```{r}
cm=c("Ptprc", "Pecam1", "Fabp4", "Apold1", "Mbp", "Dpep1", "Krt5", "Krt8", "Elf5", "Col3a1","Lum")

cg=brca[cm,]
dat= cg@assays$SCT@data %>% as.matrix() %>% t() %>% as.data.frame()
dat$cid= cg$seurat_clusters
md=aggregate(dat[,-12], by=list(dat$cid), mean)
rownames(md)=md$Group.1
```

Dendrogram representation of the expression patterns of the clusters for the given compartment genes:
```{r}
dend=hclust(dist(cor(t(md[,2:12]))), method="ward.D")
dend <- dendextend::color_branches(dend, k = 5)
dend <- dendextend::color_labels(dend, k = 5)
###
pdf(file = paste0(annotDir, "/BRCA_SCT_ClusteringSteps_1k10k.pdf"),  width = 12, height = 10)
pheatmap::pheatmap(md[,2:12], scale = "column")
pheatmap::pheatmap(cor(md[,2:12]))
pheatmap::pheatmap(cor(t(md[,2:12])))
plot(dend)
dev.off()
```

Get the cluster IDs after groupings
```{r}
res=dendextend::cutree_1k.dendrogram(dend=dend,k=5) %>% as.data.frame() #5 seems the optimal number of clusters
colnames(res)= "GroupID"
all(rownames(res)== rownames(md))
md$GroupID= res$GroupID
###
md2=aggregate(md[,-c(1,13)], by=list(md$GroupID), mean )
dat= reshape2::melt(md2, id="Group.1")
##
pdf(file = paste0(annotDir, "/Histogram_ExpressionMarkers_GroupID_RNAassay_1k10K.pdf"),  width = 12, height = 10)
ggplot(dat, aes(x=Group.1,  y=value, fill= variable))  + geom_bar(stat="identity", position=position_dodge() )+gtheme + xlab("Clusters Group ID") + ylab("Log Expression") + themplot
dev.off()
```


ANNOTATION through the top expressed gene per group:

Relevel to extract the gene name which has the max mean expression level:
```{r}
annot=dat %>%  group_by(Group.1) %>%   slice_max(value) %>% as.data.frame()

annot$type= rep(NA, nrow(annot))
 for( i in 1:nrow(annot)){
   if ( annot[i,"variable"]== "Col3a1") {
     annot$type[i]= "Fibroblast"}
    if ( annot[i,"variable"]== "Ptprc") {
      annot[i,"type"]= "Immune"}
    if ( annot[i,"variable"]== "Fabp4") {
      annot[i,"type"]= "Endothelial"}
  if ( annot[i,"variable"]== "Krt8") {
      annot[i,"type"]= "Epithelial"}
   if ( annot[i,"variable"]== "Krt5") {
     annot$type[i]= "Epithelial"}
 }
colnames(annot)=c("GroupID", colnames(annot)[2:ncol(annot)])
###
res$celltype= res$GroupID
res$seurat_clusters= rownames(res)
##
annot.final= merge(annot, res, by="GroupID")
```


Group.1 variable    value
1       1   Col3a1 3.283342
2       2     Cd52 2.174336
3       3    Fabp4 4.082335
4       4     Krt8 1.358491
5       5     Krt8 2.041874 ; as well as Krt5

Rename the clusters by their corresponding cell type compartment:
```{r}
Idents(brca)="seurat_clusters"
ct= annot.final$type
names(ct)= annot.final$seurat_clusters
##
brca=RenameIdents(brca, ct)
brca$type= Idents(brca)

rm(ct); rm(md2); rm(md); rm(dend)
rm(cm);rm(dat)
rm(cg); rm(res)
rm(annot)
rm(res)
```

##---------------------------- SET COLORS FOR ALL FIGURES ----------------------------------- ##

```{r}
#----------------- COL IS DECICATED FOR EPITHELIAL ----------------- ##
grey= c("#656565" , "#555555","#454545","#808080" , "#737373","#C8C8C8","#A4A4A4") #MT
pink=c("#D6229A", "#8D3094", "#9F79B8", "#9E2569")     #ST
#preles= "#DA308E"
preles="#ACEB0E"
normal=c("#1B9E77" ,"#4D8D58" , "#7F7C39" , "#B06C1A")
blue= c("#70B0D7","#529BCB", "#61A7D2","#1D6AAD") #LT
###
col= c(normal, preles,pink,grey,blue)
names(col)=c("Basal","Luminal H-S",  "Avd" ,"LP", "P16+ Pre-lesional", "Mki67+ ST","Cytl1+ ST", "Fgf8+ ST", "Spp1+ ST","Krt7+ MT" ,"Odc1+ MT","Col1a2+ MT","Dcn+ MT", "S100a4+ MT", "Uck2+ MT","Klf2+ MT" ,"Meis2+ LT","Clu+ LT","Mgp+ LT","Apod+ LT")
##

samplecols = colorRampPalette(brewer.pal(6, "Set1")) (6)
rm(grey); rm(pink);rm(preles);rm(normal);rm(blue)


sample= c( "#12B32A", "#0C661A"   ,     "#C483DB"   ,     "#8B1EA1"   ,     "#A69E9E"   ,     "#536BB8")
names(sample)= c("Control_Cre_2.7M",   "Control_Cre_3M", "Control_Cre_5.2M",  "T99", "T94", "T6413" )


Phase= c( "G1"="#D4C5C5", "S"= "#B0994D", "G2M"= "#F5D13F" )
type= c("Fibroblast"="#226B8F" ,      "Immune"= "#B54141", "Endothelial"="#C4C043",  "Epithelial"=  "#8983A6") 

brca_cols= list(sample, Phase, type, col)
names(brca_cols)=c("sample", "Phase", "type", "col")
## SAVE IT AS A QS FILE
qs::qsave(brca_cols, paste0(RDatadir, "/BRCA_ColorList.qs"))

```



#--------------------------START HERE IF YOU ONLY WANT TO GENERATE BRCA FIGURES-----------------------------------------#

In case you did not upload the brca seurat, here it is:

```{r}
brca=qs::qread( paste0(RDatadir, "/BRCA_SCT_1k10k.qs"))
```

#--------------------SUPP FIG1 ---------------------------#

```{r}
Idents(brca)="type"
png(paste0(annotDir, "/UMAP_BRCA_WOLegend_Type.png"), height=1000, width = 1000,res=300)
p <- DimPlot(brca, cols=brca_cols$type, label = F, repel = F, label.box = F, pt.size = 0.75)  + NoAxes() + ggtitle("") + NoLegend()
p[[1]]$layers[[1]]$aes_params$alpha = 0.4
print(p)
dev.off()

##########
 ## Save the legend ###
 leg <- ggpubr::as_ggplot(cowplot::get_legend(p))
pdf(file = paste0(annotDir, "/LEGEND_Umap_BRCA_TYPE.pdf"), width = 10, height = 10)
  print(leg)
dev.off()
############################

umap_custom <- as.data.frame(brca[["umap"]]@cell.embeddings)
metada_cells_col <- brca_cols$type[match(brca$type,names(brca_cols$type))]
draw <- sample(1:dim(umap_custom)[1])

png(paste0(annotDir,"/UMAP_BRCA_SAMPLE.png"),height=1000, width = 1000,res=300)
p <- ggplot(umap_custom[draw,], aes(x =UMAP_1, y = UMAP_2 ))+ geom_point(size = 2, shape = 21,fill = metada_cells_col[draw],alpha=1) + theme_classic() 
print(p)
dev.off()
#### GET THE LEGEND:
 p <- ggplot(umap_custom[draw,], aes(x =UMAP_1, y = UMAP_2 ,fill = metada_cells_col[draw] ))+ geom_point(size = 2, shape = 21,alpha=1) + theme_classic() + theme(legend.position = "right") + scale_fill_manual(values= brca_cols$type)

 leg <- as_ggplot(get_legend(p))
 pdf(file = paste0(annotDir, "/LEGEND_Umap_BRCA_TYPE_New.pdf"), width = 10, height = 10)
   print(leg)
 dev.off()

```

```{r}
Idents(brca)="sample"
png(paste0(annotDir, "/UMAP_BRCA_WOLegend_Sample.png"), height=1000, width = 1000,res=300)
p <- DimPlot(brca, cols=brca_cols$sample, label = F, repel = F, label.box = F, pt.size = 0.75)  + NoAxes() + ggtitle("")  + NoLegend()
p[[1]]$layers[[1]]$aes_params$alpha = 0.4
print(p)
dev.off()

##################################

##Other representation with non transparent circles, need shuffling to have a representation of all
umap_custom <- as.data.frame(brca[["umap"]]@cell.embeddings)
metada_cells_col <- brca_cols$sample[match(brca$sample,names(brca_cols$sample))]
draw <- sample(1:dim(umap_custom)[1])

png(paste0(annotDir,"/UMAP_BRCA_TYPE_Sample.png"),height=1000, width = 1000,res=300)
p <- ggplot(umap_custom[draw,], aes(x =UMAP_1, y = UMAP_2 ))+ geom_point(size = 2, shape = 21,fill = metada_cells_col[draw],alpha=1) + theme_classic() 
print(p)
dev.off()



####
 ## Save the legend ###
 leg <- as_ggplot(get_legend(p))
pdf(file = paste0(annotDir, "/LEGEND_Umap_BRCA_SAMPLE.pdf"), width = 10, height = 10)
  print(leg)
dev.off()
```


#------------------ SUPP FIG 1 --------------------------------------------#

UMAP representation of the different Markers:
```{r}
##Poruce the png figures separatelt yo make sure the figures are not stucked:
p <-FeaturePlot(brca, features =c("Ptprc","H2-Eb1", "Col3a1","Dpep1", "Pecam1","Fabp4","Apold1","Krt5","Krt8"), combine = F, pt.size = 0.75 ) 

for(i in 1:length(p)) {
  p[[i]] <- p[[i]] + NoAxes() + scale_color_viridis(option="B",limits = c(0.05, 6), na.value= "grey",direction=-1) + NoLegend() + ggtitle("")

  p[[i]][[1]]$layers[[1]]$aes_params$alpha = 0.4}

g <-cowplot::plot_grid(plotlist = p, ncol = 3)  

png(paste0(annotDir, "/Umap_BRCA_MainMarkers_WOLegend.png"), height=3000, width = 3000,res=300)
print(g)
dev.off()

```


```{r}
sg=c("Ptprc","H2-Eb1", "Col3a1","Dpep1", "Pecam1","Fabp4","Apold1","Krt5","Krt8")

for (i in seq_len(length(sg))){
    p<- better_gene_umap(brca, sg, i, 0.75) 
       png(paste0(annotDir, "/BRCA_MainMarkers_",sg[i]  ,".png"), width = 2000, height = 2000, res = 300)
          print(p)
        dev.off()
}
```

```{r}
##save the legend:
p=FeaturePlot(brca, features = "Ptprc") +  scale_color_viridis(limits = c(0.05, 6), na.value= "grey", option="B",direction=-1) 

leg <- as_ggplot(get_legend(p))
ggsave(filename =paste0(annotDir, "/Legend_FeaturePlot_BRCA_Genes.pdf"), device = "pdf", plot = g, width = 100, height = 100, limitsize = F, dpi = 300, units = "mm")
```




###############################################################################
Update the inital BRCA Seurat object with the annotations:
NOTE: IF YOU ARE JUST TRYING TO MAKE THE PLOTS NO NEED TO SAVE AGAIN
```{r}
qs::qsave(brca, paste0(RDatadir, "/BRCA_SCT_1k10k.qs"))
```
