==> Scripts_invivo_Mouse/Step3_1_CNVInference.Rmd <==
---
title: "Step3_1_CNVInference"
author: "Melissa"
date: "04/10/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
mainDir <- here::here() # the MainDir should be the root of Tumor_BRCA_P53:
source(knitr::purl(file.path(mainDir,"Scripts","global_var_current.Rmd"), quiet=TRUE))
source(knitr::purl(file.path(mainDir,"Scripts","scRNAseq_Functions.Rmd"), quiet=TRUE))
source(knitr::purl(file.path(mainDir,"Scripts","global_variables.Rmd"), quiet=TRUE))
```

####--------- THIS SECTION IS TIME CONSUMING AND ALLOWS TO RUN INFER CNV ON THE SAMPLES SEPARATELY TAKING AS REF THE BASAL CELLS ----##

#----------------- DO NOT RUN THIS SECTION, ALL RESULTS WERE STORED IN PROPRE CNV FOLDERS -----------------------####

```{r}
gene_pos=qs::qread(paste0(RDatadir,"/Gene_Positions_Mouse.qs"))
gf= gene_pos
rownames(gf)=gf$mgi_symbol
gf=gf[,-c(1,2)]
colnames(gf)=NULL
```

InferCNV on all samples seprately
```{r}
epith= qs::qread(file= paste0(RDatadir, "/Epithelial_FinalAnnotation.qs"))

Idents(epith)="subtype"
cnt= subset(epith, idents= c("Basal","Avd","LP","P16+ Pre-lesional","Luminal H-S"))
st= subset(epith, idents=c("Mki67+ ST","Cytl1+ ST",  "Fgf8+ ST","Spp1+ ST","Basal"))
mt=subset(epith, idents=c("Krt7+ MT",  "Odc1+ MT", "Col1a2+ MT", "Dcn+ MT", "S100a4+ MT", "Uck2+ MT","Klf2+ MT","Basal"))
lt= subset(epith, idents= c("Meis2+ LT","Clu+ LT","Apod+ LT","Mgp+ LT","Basal"))
```



```{r}
cnt=Wrapper_inferCNV(cnt, ref_group="Basal",outputPath=paste0(ocntcnvDir,"/"), analysis_mode="samples",HMM_type="i3")
qs::qsave(cnt, paste0(ocntcnvDir, "/CTRL_Seurat_InferCNV.qs"))
##
cnt <- qs::qread(paste0(ocntcnvDir, "/CTRL_Seurat_InferCNV.qs"))
```


```{r}
st=Wrapper_inferCNV(st, ref_group="Basal",outputPath=paste0(ostcnvDir,"/"), analysis_mode="samples",HMM_type="i3")
qs::qsave(st, paste0(ocntcnvDir, "/ST_Seurat_InferCNV.qs"))
st <- qs::qread(paste0(ocntcnvDir, "/ST_Seurat_InferCNV.qs"))
ic_st <- readRDS(paste0(ostcnvDir, "/run.final.infercnv_obj"))
st= subset(st, idents=c("Mki67+ ST","Cytl1+ ST",  "Fgf8+ ST","Spp1+ ST"))
```

```{r}
lt=Wrapper_inferCNV(lt, ref_group="Basal",outputPath=paste0(oltcnvDir,"/"), analysis_mode="samples",HMM_type="i3")
qs::qsave(lt, paste0(oltcnvDir, "/LT_Seurat_InferCNV.qs"))
```

```{r}
mt=Wrapper_inferCNV(omt, ref_group="Basal",outputPath=paste0(omtcnvDir,"/"), analysis_mode="samples",HMM_type="i3")
qs::qsave(mt, paste0(omtcnvDir, "/MT_Seurat_InferCNV.qs"))
```
